pyproject.toml
[project]
name = "django-api-boilerplate"
version = "0.1.0"
description = "A clean Django REST API boilerplate with Docker support"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "Django==5.1.4",
    "djangorestframework==3.14.0",
    "djangorestframework-simplejwt==5.5.1",
    "django-cors-headers==4.3.1",
    "celery==5.3.6",
    "redis==5.0.1",
    "gunicorn==21.2.0",
    "python-dotenv==1.0.0",
    "Pillow==10.2.0",
    "pgvector==0.2.4",
    "numpy>=1.25.2",
    "openai>=1.68.0",
    "pypdf==3.17.1",
    "tiktoken>=0.9.0",
    "psycopg2-binary==2.9.9",
]

[project.optional-dependencies]
dev = [
    "black==23.12.1",
    "flake8==7.0.0",
    "isort==5.13.2",
]
prod = [
    "psycopg2==2.9.9",  # Source for production
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["django_api_project", "core", "accounts", "chat", "rag"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_django = "django"
known_first_party = ["core", "accounts", "chat", "rag"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "DJANGO", "FIRSTPARTY", "LOCALFOLDER"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "django_api_project.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "-v --tb=short"
